mm: memcontrol: Update callback signature in mem_cgroup_scan_tasks()
Modify the callback prototype in mem_cgroup_scan_tasks() from:
    int (*fn)(struct task_struct *, void *)
to:
    int (*fn)(struct task_struct *, const char *, struct oom_control *)

This change allows passing an identifier ("memcontrol") and an OOM control
structure to the callback, providing additional context for memory control
operations and improved handling of out-of-memory situations. The invocation
is updated accordingly to pass the constant string "memcontrol".

Note: This change applies only to non-root memory cgroups.

Change-Id: I15c20a690d76342eeca527a9e132766e41f221dd
Signed-off-by: jmpfbmx <jmpf.bmx@gmail.com>